PSScriptAnalyzer fixes
Fixes to powershell scripts according to recommendations from PSScriptAnalyzer .

Most of these changes are replacing an alias for the full name (eg: "%" -> "ForEachObject")

And many of them pertain to that comparisons with $null likely should have $null on the left. Because in PS, the comparison $var -eq $null gives you the collection of $null values in $var, whereas $null -eq $var checks if $var is null.