Make disk-wipe embedded
Attempting to install the disk-wipe package gave me a 403 error.  Maybe the site is blocking scripted downloads?

Anyway, manually downloaded it and checked the included license.  It is freely distributable.  IMHO, an embedded Chocolatey package is better -- especially for old, possibly abandoned software -- to give users another option for obtaining it, so I converted this to an embedded package.

Hope you approve!